Added 64-bit platform targets and re-targeted all projects to x64 to silence compiler warnings.
Moved CubeGridEntity to its own file.
