+++ SEModAPI.sln
@@ -1,6 +1,8 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
+# Visual Studio 2013
+VisualStudioVersion = 12.0.31101.0
+MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SEModAPI", "SEModAPI\SEModAPI.csproj", "{94E6C913-9456-408D-BDEE-56B48B5397D4}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SEModAPIInternal", "SEModAPIInternal\SEModAPIInternal.csproj", "{8C4B07B6-0864-497A-ACAA-18BABE2F4168}"
@@ -9,15 +11,15 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SEModAPIExtensions", "SEMod
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SEServerExtender", "SEServerExtender\SEServerExtender.csproj", "{A442CD75-92D3-430C-84FB-889677EB1FD3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SEServerGUI", "SEServerGUI\SEServerGUI.csproj", "{F8134B84-4650-45E0-8367-F81B265E122D}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Debug|Mixed Platforms = Debug|Mixed Platforms
+		Debug|x64 = Debug|x64
 		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
 		Release|Mixed Platforms = Release|Mixed Platforms
+		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
@@ -25,54 +27,58 @@ Global
 		{94E6C913-9456-408D-BDEE-56B48B5397D4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{94E6C913-9456-408D-BDEE-56B48B5397D4}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
 		{94E6C913-9456-408D-BDEE-56B48B5397D4}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{94E6C913-9456-408D-BDEE-56B48B5397D4}.Debug|x64.ActiveCfg = Debug|x64
+		{94E6C913-9456-408D-BDEE-56B48B5397D4}.Debug|x64.Build.0 = Debug|x64
 		{94E6C913-9456-408D-BDEE-56B48B5397D4}.Debug|x86.ActiveCfg = Debug|Any CPU
 		{94E6C913-9456-408D-BDEE-56B48B5397D4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{94E6C913-9456-408D-BDEE-56B48B5397D4}.Release|Any CPU.Build.0 = Release|Any CPU
 		{94E6C913-9456-408D-BDEE-56B48B5397D4}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
 		{94E6C913-9456-408D-BDEE-56B48B5397D4}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{94E6C913-9456-408D-BDEE-56B48B5397D4}.Release|x64.ActiveCfg = Release|x64
+		{94E6C913-9456-408D-BDEE-56B48B5397D4}.Release|x64.Build.0 = Release|x64
 		{94E6C913-9456-408D-BDEE-56B48B5397D4}.Release|x86.ActiveCfg = Release|Any CPU
 		{8C4B07B6-0864-497A-ACAA-18BABE2F4168}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8C4B07B6-0864-497A-ACAA-18BABE2F4168}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8C4B07B6-0864-497A-ACAA-18BABE2F4168}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
 		{8C4B07B6-0864-497A-ACAA-18BABE2F4168}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{8C4B07B6-0864-497A-ACAA-18BABE2F4168}.Debug|x64.ActiveCfg = Debug|x64
+		{8C4B07B6-0864-497A-ACAA-18BABE2F4168}.Debug|x64.Build.0 = Debug|x64
 		{8C4B07B6-0864-497A-ACAA-18BABE2F4168}.Debug|x86.ActiveCfg = Debug|Any CPU
 		{8C4B07B6-0864-497A-ACAA-18BABE2F4168}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8C4B07B6-0864-497A-ACAA-18BABE2F4168}.Release|Any CPU.Build.0 = Release|Any CPU
 		{8C4B07B6-0864-497A-ACAA-18BABE2F4168}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
 		{8C4B07B6-0864-497A-ACAA-18BABE2F4168}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{8C4B07B6-0864-497A-ACAA-18BABE2F4168}.Release|x64.ActiveCfg = Release|x64
+		{8C4B07B6-0864-497A-ACAA-18BABE2F4168}.Release|x64.Build.0 = Release|x64
 		{8C4B07B6-0864-497A-ACAA-18BABE2F4168}.Release|x86.ActiveCfg = Release|Any CPU
 		{727B1141-2124-4A31-82FA-CAE218501215}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{727B1141-2124-4A31-82FA-CAE218501215}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{727B1141-2124-4A31-82FA-CAE218501215}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
 		{727B1141-2124-4A31-82FA-CAE218501215}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{727B1141-2124-4A31-82FA-CAE218501215}.Debug|x64.ActiveCfg = Debug|x64
+		{727B1141-2124-4A31-82FA-CAE218501215}.Debug|x64.Build.0 = Debug|x64
 		{727B1141-2124-4A31-82FA-CAE218501215}.Debug|x86.ActiveCfg = Debug|Any CPU
 		{727B1141-2124-4A31-82FA-CAE218501215}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{727B1141-2124-4A31-82FA-CAE218501215}.Release|Any CPU.Build.0 = Release|Any CPU
 		{727B1141-2124-4A31-82FA-CAE218501215}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
 		{727B1141-2124-4A31-82FA-CAE218501215}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{727B1141-2124-4A31-82FA-CAE218501215}.Release|x64.ActiveCfg = Release|x64
+		{727B1141-2124-4A31-82FA-CAE218501215}.Release|x64.Build.0 = Release|x64
 		{727B1141-2124-4A31-82FA-CAE218501215}.Release|x86.ActiveCfg = Release|Any CPU
 		{A442CD75-92D3-430C-84FB-889677EB1FD3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A442CD75-92D3-430C-84FB-889677EB1FD3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A442CD75-92D3-430C-84FB-889677EB1FD3}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
 		{A442CD75-92D3-430C-84FB-889677EB1FD3}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{A442CD75-92D3-430C-84FB-889677EB1FD3}.Debug|x64.ActiveCfg = Debug|x64
+		{A442CD75-92D3-430C-84FB-889677EB1FD3}.Debug|x64.Build.0 = Debug|x64
 		{A442CD75-92D3-430C-84FB-889677EB1FD3}.Debug|x86.ActiveCfg = Debug|Any CPU
 		{A442CD75-92D3-430C-84FB-889677EB1FD3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A442CD75-92D3-430C-84FB-889677EB1FD3}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A442CD75-92D3-430C-84FB-889677EB1FD3}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
 		{A442CD75-92D3-430C-84FB-889677EB1FD3}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{A442CD75-92D3-430C-84FB-889677EB1FD3}.Release|x64.ActiveCfg = Release|x64
+		{A442CD75-92D3-430C-84FB-889677EB1FD3}.Release|x64.Build.0 = Release|x64
 		{A442CD75-92D3-430C-84FB-889677EB1FD3}.Release|x86.ActiveCfg = Release|Any CPU
-		{F8134B84-4650-45E0-8367-F81B265E122D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F8134B84-4650-45E0-8367-F81B265E122D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F8134B84-4650-45E0-8367-F81B265E122D}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{F8134B84-4650-45E0-8367-F81B265E122D}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{F8134B84-4650-45E0-8367-F81B265E122D}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{F8134B84-4650-45E0-8367-F81B265E122D}.Debug|x86.Build.0 = Debug|Any CPU
-		{F8134B84-4650-45E0-8367-F81B265E122D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F8134B84-4650-45E0-8367-F81B265E122D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F8134B84-4650-45E0-8367-F81B265E122D}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{F8134B84-4650-45E0-8367-F81B265E122D}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{F8134B84-4650-45E0-8367-F81B265E122D}.Release|x86.ActiveCfg = Release|Any CPU
-		{F8134B84-4650-45E0-8367-F81B265E122D}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
