+++ SEServerExtender/SEServerExtender.csproj
@@ -57,6 +57,25 @@
     <StartupObject>
     </StartupObject>
   </PropertyGroup>
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|x64'">
+    <DebugSymbols>true</DebugSymbols>
+    <OutputPath>bin\x64\Debug\</OutputPath>
+    <DefineConstants>DEBUG;TRACE</DefineConstants>
+    <WarningLevel>3</WarningLevel>
+    <DebugType>full</DebugType>
+    <PlatformTarget>x64</PlatformTarget>
+    <ErrorReport>prompt</ErrorReport>
+    <CodeAnalysisRuleSet>AllRules.ruleset</CodeAnalysisRuleSet>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Release|x64'">
+    <OutputPath>bin\x64\Release\</OutputPath>
+    <DefineConstants>TRACE</DefineConstants>
+    <Optimize>true</Optimize>
+    <DebugType>pdbonly</DebugType>
+    <PlatformTarget>x64</PlatformTarget>
+    <ErrorReport>prompt</ErrorReport>
+    <CodeAnalysisRuleSet>AllRules.ruleset</CodeAnalysisRuleSet>
+  </PropertyGroup>
   <ItemGroup>
     <Reference Include="HavokWrapper, Version=1.0.5326.36939, Culture=neutral, processorArchitecture=AMD64">
       <SpecificVersion>False</SpecificVersion>
@@ -126,20 +145,6 @@
     </Compile>
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\SEModAPIExtensions\SEModAPIExtensions.csproj">
-      <Project>{727B1141-2124-4A31-82FA-CAE218501215}</Project>
-      <Name>SEModAPIExtensions</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\SEModAPIInternal\SEModAPIInternal.csproj">
-      <Project>{8C4B07B6-0864-497A-ACAA-18BABE2F4168}</Project>
-      <Name>SEModAPIInternal</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\SEModAPI\SEModAPI.csproj">
-      <Project>{94E6C913-9456-408D-BDEE-56B48B5397D4}</Project>
-      <Name>SEModAPI</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
     <EmbeddedResource Include="CubeBlockDialog.resx">
       <DependentUpon>CubeBlockDialog.cs</DependentUpon>
     </EmbeddedResource>
@@ -180,6 +185,20 @@
     <Folder Include="InternalService\" />
     <Folder Include="PluginService\" />
     <Folder Include="ServerService\" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\SEModAPIExtensions\SEModAPIExtensions.csproj">
+      <Project>{727b1141-2124-4a31-82fa-cae218501215}</Project>
+      <Name>SEModAPIExtensions</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\SEModAPIInternal\SEModAPIInternal.csproj">
+      <Project>{8c4b07b6-0864-497a-acaa-18babe2f4168}</Project>
+      <Name>SEModAPIInternal</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\SEModAPI\SEModAPI.csproj">
+      <Project>{94e6c913-9456-408d-bdee-56b48b5397d4}</Project>
+      <Name>SEModAPI</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
